Heart framework created
Rails 7 server with postgres db, using stimulus in the front end.
Initial structure is to have device models that keep information about various devices. As well as cards that keep configuration information about what is displayed on the dashboard.
Added capistrano for deploying and set it all up on docker to make admin easy.